Added Vertex Light Support to Flat Lit Toon
This will add working vertex light handling to the Flat Lit Toon shader.
Computes vertex light contribution in the vert function and properly
carries it to the frag function.

I'm unsure exactly how this contribution should be added but it seemed
like adding it to the direction contribution was desired. Light works
additive, sometimes right?

Anyway, it's based partially on the source of the Standard Shader's
forwardbase vert function that handles vertex lighting.